+++ lib/language_server/protocol/transport/stdio/reader.rb
@@ -1,36 +1,10 @@
-require "json"
-
 module LanguageServer
   module Protocol
     module Transport
       module Stdio
-        class Reader
-          def read(&block)
-            buffer = ""
-            header_parsed = false
-            content_length = nil
-
-            while char = STDIN.getc
-              buffer << char
-
-              unless header_parsed
-                if buffer[-4..-1] == "\r\n" * 2
-                  content_length = buffer.match(/Content-Length: (\d+)/i)[1].to_i
-
-                  header_parsed = true
-                  buffer.clear
-                end
-              else
-                if buffer.bytesize == content_length
-                  request = JSON.parse(buffer, symbolize_names: true)
-
-                  block.call(request)
-
-                  header_parsed = false
-                  buffer.clear
-                end
-              end
-            end
+        class Reader < Io::Reader
+          def initialize
+            super STDIN
           end
         end
       end